restructuring dir for submodule
