--- plynx/cdp/handlers/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2022, N3 Hub Limited
-# All rights reserved.
-
-from plynx.cdp.handlers.webhook_handler import StreamDataHandler, RunWorkflowHandler
