+++ plynx/cdp/src/handlers/base.py
@@ -1,14 +1,15 @@
 # Copyright (C) 2022, N3 Hub Limited
 # All rights reserved.
 
+import logging
 from abc import ABC, abstractmethod
 
-from plynx.cdp import get_logger
 
 class BaseHandler(ABC):
+    """Base event handler."""
 
     def __init__(self):
-        self.logger = get_logger()
+        self.logger = logging.getLogger(self.__class__.__name__)
 
     @abstractmethod
     def process(self):
