+++ plynx/cdp/src/handlers/webhook_handler.py
@@ -3,9 +3,9 @@
 
 from typing import Any, List
 
-from plynx.cdp.handlers.base import BaseHandler
-from plynx.cdp.data import EventMessage, StreamData, WorkflowNode, OperationNode
-from plynx.cdp.constants import WebhookEvents, NodeTypes, DataSource, StreamDataFields
+from plynx.cdp.src.handlers.base import BaseHandler
+from plynx.cdp.src.data import EventMessage, StreamData, WorkflowNode, OperationNode
+from plynx.cdp.src.constants import WebhookEvents, NodeTypes, DataSource, StreamDataFields
 from plynx.db.node import Node, Parameter, NodeClonePolicy
 from plynx.utils.db_connector import get_db_connector
 from plynx.constants import Collections, NodeStatus
