+++ plynx/cdp/src/operators/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (C) 2022, N3 Hub Limited
 # All rights reserved.
 
-from plynx.cdp.operators.python import GROUP as cdp_python_group
+from plynx.cdp.src.operators.python import GROUP as cdp_python_group
 
 
 COLLECTION = [cdp_python_group]
