+++ plynx/cdp/src/utils/validators.py
@@ -3,5 +3,7 @@
 
 
 def check_empty(value):
+    """Check if the given value was empty."""
+
     assert value, "Value cannot be empty."
     return value
